int32 instead of int64

+++ cuda/src/kernels/array/broadcast.rs
@@ -113,9 +113,9 @@ impl MultiBroadcast {
 
         launch_args.set_view(&i_view);
         launch_args.set_view(&o_view);
-        launch_args.set_slice::<i64>(&input_broadcast_strides);
-        launch_args.set_slice::<i64>(out_shape);
-        launch_args.set_slice::<i64>(output.strides());
+        launch_args.set_slice::<i32>(&input_broadcast_strides);
+        launch_args.set_slice::<i32>(out_shape);
+        launch_args.set_slice::<i32>(output.strides());
 
         let cfg = utils::cuda_launch_cfg_for_cpy(out_shape);
